# More efficiently validate NumericalDataSeries