`_validate` currently loops through the entire data array in an `Array.prototype.every` callback, repeating the same work `data.length` times.