@alamb 2017-01-18
I am curious -- did  you do any measurements to show this code was slow / could be faster, or was this change based on code inspection?