@nawatts 2017-01-18
@alamb Just code inspection before changing. Since you mentioned it, I put together a quick benchmark. For validating a valid array of data points with `x_value` and `y_value` keys (times and speedup are average of 100 trials):

| # data points | old implementation | new implementation | speedup |
| --- | --- | --- | --- |
100 | 10.75 ms | 0.13 ms | 83x |
1000 | 1136 ms | 1.12 ms | 1014x |